requirements.txt
python-telegram-bot[job-queue]==22.5
httpx==0.28.1
aiosqlite==0.21.0
python-dotenv==1.1.0
openpyxl==3.1.5
